// web/src/api/models/TypesLLDPNeighbor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesLLDPNeighbor
 */
export interface TypesLLDPNeighbor {
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    age?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    capabilities?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    chassis_id?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    chassis_name?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    local_iface: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    mgmt_ip?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    port_descr?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    port_id?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    protocol: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    sys_descr?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLLDPNeighbor
     */
    vlan?: string;
}

/**
 * Check if a given object implements the TypesLLDPNeighbor interface.
 */
export function instanceOfTypesLLDPNeighbor(value: object): value is TypesLLDPNeighbor {
    if (!('local_iface' in value) || value['local_iface'] === undefined) return false;
    if (!('protocol' in value) || value['protocol'] === undefined) return false;
    return true;
}

export function TypesLLDPNeighborFromJSON(json: any): TypesLLDPNeighbor {
    return TypesLLDPNeighborFromJSONTyped(json, false);
}

export function TypesLLDPNeighborFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesLLDPNeighbor {
    if (json == null) {
        return json;
    }
    return {
        
        'age': json['age'] == null ? undefined : json['age'],
        'capabilities': json['capabilities'] == null ? undefined : json['capabilities'],
        'chassis_id': json['chassis_id'] == null ? undefined : json['chassis_id'],
        'chassis_name': json['chassis_name'] == null ? undefined : json['chassis_name'],
        'local_iface': json['local_iface'],
        'mgmt_ip': json['mgmt_ip'] == null ? undefined : json['mgmt_ip'],
        'port_descr': json['port_descr'] == null ? undefined : json['port_descr'],
        'port_id': json['port_id'] == null ? undefined : json['port_id'],
        'protocol': json['protocol'],
        'sys_descr': json['sys_descr'] == null ? undefined : json['sys_descr'],
        'vlan': json['vlan'] == null ? undefined : json['vlan'],
    };
}

export function TypesLLDPNeighborToJSON(json: any): TypesLLDPNeighbor {
    return TypesLLDPNeighborToJSONTyped(json, false);
}

export function TypesLLDPNeighborToJSONTyped(value?: TypesLLDPNeighbor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'age': value['age'],
        'capabilities': value['capabilities'],
        'chassis_id': value['chassis_id'],
        'chassis_name': value['chassis_name'],
        'local_iface': value['local_iface'],
        'mgmt_ip': value['mgmt_ip'],
        'port_descr': value['port_descr'],
        'port_id': value['port_id'],
        'protocol': value['protocol'],
        'sys_descr': value['sys_descr'],
        'vlan': value['vlan'],
    };
}